코드 수정 : 보안 관련 추가

+++ delete.php
@@ -2,10 +2,15 @@
 <?php
 require 'db.php';
 
-$id = $_GET['id'] ?? 0;
+$id =(int)$_POST['id'] ?? 0;
 
-$stmt = $pdo->prepare("DELETE FROM todos WHERE id = ?");
-$stmt->execute([$id]);
+if ($id > 0) {
+    $stmt = $pdo->prepare("DELETE FROM todos WHERE id = ?");
+    $stmt->execute([$id]);
+
+    if ($stmt->rowCount() === 0) {
+    }
+}
 
 header("Location: index.php");
 exit;
